datatype has to be long
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@23504 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/services/networking/wshaper.c
@@ -75,7 +75,7 @@ extern char *get_tcfmark(uint32 mark);
 #endif
 
 #ifdef HAVE_AQOS
-extern void add_client_classes(unsigned int base, unsigned int uprate, unsigned int downrate, unsigned int lanrate, unsigned int level);
+extern void add_client_classes(unsigned int base, unsigned int uprate, unsigned int downrate, unsigned long lanrate, unsigned int level);
 #else
 extern void add_client_classes(unsigned int base, unsigned int level);
 #endif
